backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
alembic==1.12.1
